+++ CHANGES.md
@@ -1,7 +1,8 @@
 ## Pending
 
 * Use arc4random_buf instead of getrandom on Android before getrandom
-  became available in API 28.
+  became available in API 28 (#259 @jonahbeckford)
+* Define fill_bytes for MSVC (#259 @jonahbeckford)
 
 ## v2.0.0 (2025-02-03)